requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
gunicorn
psycopg2-binary
djangorestframework~=3.12.4